a bloody, sweaty, tear-filled refactor of triad rendering is complete

+++ src/dray/rendering/annotator.hpp
@@ -31,7 +31,7 @@ public:
                       const std::vector<std::string> &field_names,
                       std::vector<ColorMap> &color_maps);
   void draw_triad(Framebuffer &fb,
-				  Vec<int32, 2> pos, // screen space coords where we want the triad to be centered
+				  Vec<float32, 2> pos, // screen space coords where we want the triad to be centered (-1 1)
 				  float32 distance,
 				  Camera &camera);
